+++ App.js
@@ -1,21 +1,27 @@
 import { StatusBar } from 'expo-status-bar';
 import React from 'react';
+import { render } from 'react-dom';
 import { StyleSheet, Text, View } from 'react-native';
+import {createStackNavigator, createAppContainer } from "react-navigation";
+import Home from './screens/Home';
 
-export default function App() {
-  return (
-    <View style={styles.container}>
-      <Text>Open up App.js to start working on your app!</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
-}
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
+const MainNavigator = createStackNavigator({
+  Home: {screen: Home}
+}, {
+  defaultNavigationOptions: {
+    headerTintColor: "#fff",
+    headerStyle: {
+      backgroundColor: "#000000"
+    },
+    headerTitleStyle: {
+      color: "#FFF"
+    }
+  }
 });
+
+
+const App = createAppContainer(MainNavigator);
+export default App;
+
+
